# Reduce reliance on lodash methods