--- a/src/indent.js
+++ b/src/indent.js
@@ -1,5 +1,4 @@
 import through2 from 'through2';
-import _ from 'lodash';
 
 export default function Indent() {
   const NEWLINE = '\n';
@@ -18,7 +17,7 @@ export default function Indent() {
       if (indent) {
         if (preceededNewLine && !beginsNewLine) content = indent + content;
 
-        content = _.replace(content, /\n(?!\s|$)/g, `\n${indent}`);
+        content = content.replace(/\n(?!\s|$)/g, `\n${indent}`);
         inputBuffer[1].content = content;
       }