Aligning with https://github.com/jamesramsay/through2-get/pull/118 this PR does two things

- Use the per-method imports for lodash 
- Make use of native Node.js functionality when that's easily possible

Use Node.js 14.15.1 the benchmark results show no performance impact or even a tiny improvement:

main:
```
> ./test/benchmark.js

hercule#transcludeFile x 1,284 ops/sec ±6.85% (71 runs sampled)
hercule#TranscludeStream x 1,525 ops/sec ±0.79% (82 runs sampled)
hercule#transcludeString x 1,616 ops/sec ±0.68% (85 runs sampled)
```

This branch:
```
> ./test/benchmark.js

hercule#transcludeFile x 1,309 ops/sec ±8.13% (71 runs sampled)
hercule#TranscludeStream x 1,543 ops/sec ±0.93% (79 runs sampled)
hercule#transcludeString x 1,639 ops/sec ±1.25% (82 runs sampled)
```

Outside of tests the following lodash functions are still used:

- `isString`
- `uniqBy`
- `isPlainObject`